docs : code splitting
