+++ packages/assignment-4/src/advanced/main.advanced.js
@@ -1,3 +1,9 @@
-function main() {}
+import { createCartView } from './createCartView.js';
+import { createShoppingCart } from './createShoppingCart.js';
+
+function main() {
+  const cart = createShoppingCart();
+  createCartView(cart);
+}
 
 main();
